Réorganisation du code, car ça faisait lourd

+++ plugin/src/main/java/fr/euphyllia/skyllia/api/InterneAPI.java
@@ -16,7 +16,6 @@ import fr.euphyllia.skyllia.sgbd.DatabaseLoader;
 import fr.euphyllia.skyllia.sgbd.MariaDB;
 import fr.euphyllia.skyllia.sgbd.exceptions.DatabaseException;
 import net.kyori.adventure.text.minimessage.MiniMessage;
-import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.bukkit.Bukkit;
@@ -27,74 +26,121 @@ import org.jetbrains.annotations.Nullable;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.nio.file.FileSystem;
 import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
+/**
+ * Internal API class, managing NMS, database connections, and caching.
+ */
 public class InterneAPI {
 
-    private final Logger logger = LogManager.getLogger(this);
+    private static final Logger LOGGER = LogManager.getLogger(InterneAPI.class);
+
     private final Main plugin;
     private final SkyblockManager skyblockManager;
     private final CacheManager cacheManager;
-    private @Nullable DatabaseLoader database;
-    private DatabaseLoader databaseLoader;
-    private Managers managers;
+
     private WorldNMS worldNMS;
     private PlayerNMS playerNMS;
     private BiomesImpl biomesImpl;
 
+    private DatabaseLoader database;
+    private Managers managers;
+
+    /**
+     * Constructor that checks the Minecraft version and initializes NMS classes.
+     *
+     * @param plugin The main plugin instance
+     * @throws UnsupportedMinecraftVersionException if server version is not supported
+     */
     public InterneAPI(Main plugin) throws UnsupportedMinecraftVersionException {
         this.plugin = plugin;
-        this.setVersionNMS();
-        this.skyblockManager = new SkyblockManager(this.plugin);
+        setVersionNMS();
+        this.skyblockManager = new SkyblockManager(plugin);
         this.cacheManager = new CacheManager(this.skyblockManager, this);
+        loadAPI();
     }
 
-    public @Nullable DatabaseLoader getDatabaseLoader() {
-        return this.database;
-    }
-
-    public Managers getManagers() {
-        return managers;
-    }
-
-    public void setManagers(Managers managers) {
-        this.managers = managers;
-    }
-
-    private @Nullable File checkFileExist(File dataFolder, String fileName) throws IOException {
-        if (!dataFolder.exists() && (!dataFolder.mkdir())) {
-            logger.log(Level.FATAL, "Unable to create the configuration folder.");
-            return null;
-
+    /**
+     * Determines the version of NMS to use based on the Bukkit server version.
+     *
+     * @throws UnsupportedMinecraftVersionException if the version is not supported
+     */
+    private void setVersionNMS() throws UnsupportedMinecraftVersionException {
+        final String[] bukkitVersion = Bukkit.getServer().getBukkitVersion().split("-");
+        switch (bukkitVersion[0]) {
+            case "1.20", "1.20.1" -> {
+                this.worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R1.WorldNMS();
+                this.playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R1.PlayerNMS();
+                this.biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_20_R1.BiomeNMS();
+            }
+            case "1.20.2" -> {
+                this.worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R2.WorldNMS();
+                this.playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R2.PlayerNMS();
+                this.biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_20_R2.BiomeNMS();
+            }
+            case "1.20.3", "1.20.4" -> {
+                this.worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R3.WorldNMS();
+                this.playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R3.PlayerNMS();
+                this.biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_20_R3.BiomeNMS();
+            }
+            case "1.20.5", "1.20.6" -> {
+                this.worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R4.WorldNMS();
+                this.playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R4.PlayerNMS();
+                this.biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_20_R4.BiomeNMS();
+            }
+            case "1.21", "1.21.1" -> {
+                this.worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_21_R1.WorldNMS();
+                this.playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_21_R1.PlayerNMS();
+                this.biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_21_R1.BiomeNMS();
+            }
+            case "1.21.2", "1.21.3" -> {
+                this.worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_21_R2.WorldNMS();
+                this.playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_21_R2.PlayerNMS();
+                this.biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_21_R2.BiomeNMS();
+            }
+            case "1.21.4" -> {
+                this.worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_21_R3.WorldNMS();
+                this.playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_21_R3.PlayerNMS();
+                this.biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_21_R3.BiomeNMS();
+            }
+            default -> {
+                throw new UnsupportedMinecraftVersionException("Version " + bukkitVersion[0] + " not supported!");
+            }
         }
-        FileSystem fs = FileSystems.getDefault();
-        Path configFolder = fs.getPath(dataFolder.getAbsolutePath());
-
-        File configFile = new File(configFolder + File.separator + fileName);
-        if (!configFile.exists()) {
-            configFile.createNewFile();
-        }
-        return configFile;
     }
 
+
+    /**
+     * Loads a config file if it doesn't exist, then initializes it via a ConfigInitializer.
+     *
+     * @param dataFolder Path to the plugin's data folder
+     * @param fileName   The config file name
+     * @param initializer The ConfigInitializer functional interface
+     * @return true if successful, false otherwise
+     * @throws IOException if file creation fails
+     */
     public boolean setupConfigs(File dataFolder, String fileName, ConfigInitializer initializer) throws IOException {
-        File configFile = this.checkFileExist(dataFolder, fileName);
+        File configFile = checkFileExist(dataFolder, fileName);
         if (configFile == null) {
             return false;
         }
-
         try {
             initializer.initConfig(configFile);
         } catch (Exception ex) {
-            logger.log(Level.FATAL, ex.getMessage(), ex);
+            LOGGER.error("Error while initializing config: ", ex);
             return false;
         }
         return true;
     }
 
+    /**
+     * Copies a default schematic if it does not exist already.
+     *
+     * @param dataFolder The plugin's data folder
+     * @param resource   The InputStream for the default schematic
+     */
     public void setupFirstSchematic(@NotNull File dataFolder, @Nullable InputStream resource) {
         File schematicsDir = new File(dataFolder, "schematics");
         File defaultSchem = new File(schematicsDir, "default.schem");
@@ -103,19 +149,21 @@ public class InterneAPI {
         } else {
             return;
         }
-        if (!defaultSchem.exists()) {
-            try {
-                try (InputStream in = resource) {
-                    if (in != null) {
-                        Files.copy(in, defaultSchem.toPath());
-                    }
-                }
+        if (!defaultSchem.exists() && resource != null) {
+            try (InputStream in = resource) {
+                Files.copy(in, defaultSchem.toPath());
             } catch (IOException e) {
                 e.printStackTrace();
             }
         }
     }
 
+    /**
+     * Initializes the SGBD (e.g., MariaDB) from the config, creating the database schema if necessary.
+     *
+     * @return true if successful, false otherwise
+     * @throws DatabaseException if database initialization fails
+     */
     public boolean setupSGBD() throws DatabaseException {
         if (ConfigToml.mariaDBConfig != null) {
             MariaDB mariaDB = new MariaDB(ConfigToml.mariaDBConfig);
@@ -129,12 +177,68 @@ public class InterneAPI {
         }
     }
 
+    /**
+     * Returns an IslandQuery instance to access or modify island data.
+     *
+     * @return A new IslandQuery instance
+     */
     public IslandQuery getIslandQuery() {
         return new IslandQuery(this, ConfigToml.mariaDBConfig.database());
     }
 
+    /**
+     * Periodically updates the cache for a given player.
+     *
+     * @param player The player whose cache is updated
+     */
+    public void updateCache(Player player) {
+        this.cacheManager.updateCache(skyblockManager, player);
+    }
+
+    /**
+     * Loads the public API implementation.
+     */
+    private void loadAPI() {
+        fr.euphyllia.skyllia.api.SkylliaAPI.setImplementation(this.plugin, new APISkyllia(this));
+    }
+
+    /**
+     * Ensures a config file exists in the dataFolder, creating it if necessary.
+     *
+     * @param dataFolder The plugin data folder
+     * @param fileName   The file name
+     * @return The File object or null if creation fails
+     * @throws IOException if file creation fails
+     */
+    private @Nullable File checkFileExist(File dataFolder, String fileName) throws IOException {
+        if (!dataFolder.exists() && !dataFolder.mkdir()) {
+            LOGGER.error("Could not create directory: {}", dataFolder.getAbsolutePath());
+            return null;
+        }
+        Path configFolder = FileSystems.getDefault().getPath(dataFolder.getAbsolutePath());
+        File configFile = new File(configFolder + File.separator + fileName);
+        if (!configFile.exists()) {
+            configFile.createNewFile();
+        }
+        return configFile;
+    }
+
+    /* Getters and setters */
+
     public Main getPlugin() {
         return this.plugin;
+    }
+
+    public Managers getManagers() {
+        return managers;
+    }
+
+    public void setManagers(Managers managers) {
+        this.managers = managers;
+    }
+
+    public DatabaseLoader getDatabaseLoader() {
+        return this.database;
     }
 
     public SkyblockManager getSkyblockManager() {
@@ -145,55 +249,8 @@ public class InterneAPI {
         return MiniMessage.miniMessage();
     }
 
-    public void updateCache(Player player) {
-        this.cacheManager.updateCache(skyblockManager, player);
-    }
-
     public CacheManager getCacheManager() {
         return this.cacheManager;
-    }
-
-    private void setVersionNMS() throws UnsupportedMinecraftVersionException {
-        final String[] bukkitVersion = Bukkit.getServer().getBukkitVersion().split("-");
-        switch (bukkitVersion[0]) {
-            case "1.20", "1.20.1" -> {
-                worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R1.WorldNMS();
-                playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R1.PlayerNMS();
-                biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_20_R1.BiomeNMS();
-            }
-            case "1.20.2" -> {
-                worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R2.WorldNMS();
-                playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R2.PlayerNMS();
-                biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_20_R2.BiomeNMS();
-            }
-            case "1.20.3", "1.20.4" -> {
-                worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R3.WorldNMS();
-                playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R3.PlayerNMS();
-                biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_20_R3.BiomeNMS();
-            }
-            case "1.20.5", "1.20.6" -> {
-                worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R4.WorldNMS();
-                playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_20_R4.PlayerNMS();
-                biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_20_R4.BiomeNMS();
-            }
-            case "1.21", "1.21.1" -> {
-                worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_21_R1.WorldNMS();
-                playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_21_R1.PlayerNMS();
-                biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_21_R1.BiomeNMS();
-            }
-            case "1.21.2", "1.21.3" -> {
-                worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_21_R2.WorldNMS();
-                playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_21_R2.PlayerNMS();
-                biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_21_R2.BiomeNMS();
-            }
-            case "1.21.4" -> {
-                worldNMS = new fr.euphyllia.skyllia.utils.nms.v1_21_R3.WorldNMS();
-                playerNMS = new fr.euphyllia.skyllia.utils.nms.v1_21_R3.PlayerNMS();
-                biomesImpl = new fr.euphyllia.skyllia.utils.nms.v1_21_R3.BiomeNMS();
-            }
-            default ->
-                    throw new UnsupportedMinecraftVersionException("Version %s not supported !".formatted(bukkitVersion[0]));
-        }
     }
 
     public WorldNMS getWorldNMS() {
@@ -202,10 +259,6 @@ public class InterneAPI {
 
     public PlayerNMS getPlayerNMS() {
         return this.playerNMS;
-    }
-
-    public void loadAPI() {
-        SkylliaAPI.setImplementation(this.plugin, new APISkyllia(this));
     }
 
     public BiomesImpl getBiomesImpl() {
